+++ packages/react-resizable-panels-website/playwright.config.ts
@@ -2,6 +2,12 @@ import type { PlaywrightTestConfig } from "@playwright/test";
 
 const config: PlaywrightTestConfig = {
   use: {
+    // Uncomment to test in headless Chromium:
+    // launchOptions: {
+    //   slowMo: 100,
+    // },
+    // browserName: "chromium",
+    // headless: false,
     headless: true,
     viewport: { width: 1280, height: 720 },
     ignoreHTTPSErrors: true,