+++ packages/react-resizable-panels/src/utils/group.ts
@@ -83,6 +83,52 @@ export function adjustByDelta(
   return nextSizes;
 }
 
+export function callPanelCallbacks(
+  panelsArray: PanelData[],
+  prevSizes: number[],
+  nextSizes: number[]
+) {
+  nextSizes.forEach((nextSize, index) => {
+    const prevSize = prevSizes[index];
+    if (prevSize !== nextSize) {
+      const { callbacksRef } = panelsArray[index];
+      const { onCollapse, onResize } = callbacksRef.current;
+
+      if (onResize) {
+        onResize(nextSize);
+      }
+
+      if (onCollapse) {
+        if (prevSize === 0 && nextSize !== 0) {
+          onCollapse(false);
+        } else if (prevSize !== 0 && nextSize === 0) {
+          onCollapse(true);
+        }
+      }
+    }
+  });
+}
+
+export function getBeforeAndAfterIds(
+  id: string,
+  panelsArray: PanelData[]
+): [idBefore: string | null, idAFter: string | null] {
+  if (panelsArray.length < 2) {
+    return [null, null];
+  }
+
+  const index = panelsArray.findIndex((panel) => panel.id === id);
+  if (index < 0) {
+    return [null, null];
+  }
+
+  const isLastPanel = index === panelsArray.length - 1;
+  const idBefore = isLastPanel ? panelsArray[index - 1].id : id;
+  const idAfter = isLastPanel ? id : panelsArray[index + 1].id;
+
+  return [idBefore, idAfter];
+}
+
 // This method returns a number between 1 and 100 representing
 // the % of the group's overall space this panel should occupy.
 export function getFlexGrow(