+++ packages/react-resizable-panels/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.0.29
+* [#58](https://github.com/bvaughn/react-resizable-panels/pull/58): Add imperative `collapse`, `expand`, and `resize` methods to `Panel`.
 * [#57](https://github.com/bvaughn/react-resizable-panels/pull/57): Improve server rendering check to include `window.document`. This more closely matches React's own check and avoids false positives for environments that alias `window` to some global object.
 
 ## 0.0.28